+++ src/main/java/org/example/logisticapplication/domain/Distance/DistanceDto.java
@@ -6,10 +6,10 @@ public record DistanceDto(
         @Null
         Long id,
 
-        Integer fromCityId,
+        Long fromCityId,
 
-        Integer toCityId,
+        Long toCityId,
 
-        Integer distance
+        Double distance
 ) {
 }
